max_new_tokens as a parameter to answer_question and default 256
max_new_tokens is added as a parameter to answer_question() and default is set to 256 instead of 512. Reason for the latter: in the test generations I found about 256 is enough. Also in order to reduce the delay when the model enters in loops generating .............. and taking a lot of time (can an interrupt be added when capturing a repetition of tokens, repetition_penalty etc.? )